Styling

+++ src/utils/useFavorites.ts
@@ -3,13 +3,13 @@ import { useEffect, useState } from 'react';
 import { BASE_URL } from '../api/baseApi';
 import { LOGIN_PATH } from '../routes/consts';
 import httpClient from '../api/httpClient';
+import { toast } from 'react-hot-toast';
 import { useAuthContext } from '../hooks/useAuthContext';
 import { useNavigate } from 'react-router-dom';
 import { useUserFavorites } from '../hooks/useUserFavorites';
 
 const useFavorites = () => {
   const [likedPhotos, setLikedPhotos] = useState<number[]>([]);
-  //const initialized = useRef(false);
   const { user, isAuthenticated } = useAuthContext();
   const { fetchFavorites, favorites } = useUserFavorites();
   const navigate = useNavigate();
@@ -25,24 +25,33 @@ const useFavorites = () => {
           { withCredentials: true },
         );
         fetchFavorites();
+        toast.success('Favorite image added!');
       } catch (error) {
         console.error('Error adding favorite', error);
+        toast.error('Could not add to favorites');
       }
     } else {
       console.log('User not authenticated, cannot add favorite.');
       navigate(LOGIN_PATH);
+      toast.error('Please log in to add favorites');
     }
   };
 
   const deleteFavorite = async (itemId: number, fetchFavorites: () => void) => {
     if (isAuthenticated && user) {
-      try {
-        await httpClient.delete(`${BASE_URL}users/favorites/${itemId}`, { withCredentials: true });
-        //setLikedPhotos((prevLikedPhotos) => prevLikedPhotos.filter((id) => id !== itemId));
+      const confirmed = window.confirm('Are you sure you want to remove this image?');
+      if (confirmed) {
+        try {
+          await httpClient.delete(`${BASE_URL}users/favorites/${itemId}`, { withCredentials: true });
 
-        fetchFavorites();
-      } catch (error) {
-        console.error('Error deleting favorite', error);
+          fetchFavorites();
+          toast.success('Image removed');
+        } catch (error) {
+          console.error('Error deleting favorite', error);
+          toast.error('Error removing favorite');
+        }
+      } else {
+        return;
       }
     }
   };
